a lot more test and bugfixes, ignore derived attributes in rtext language by default

+++ lib/rtext/frontend/connector.rb
@@ -9,6 +9,8 @@ class Connector
 include Process
 include RText::MessageHelper
 
+ConnectTimeout = 10
+
 def initialize(config, options={})
   @config = config
   @logger = options[:logger]
@@ -16,10 +18,15 @@ def initialize(config, options={})
   @invocation_id = 1
   @invocations = {}
   @busy = false
+  @busy_start_time = nil
   @connection_listener = options[:connect_callback]
+  @outfile_provider = options[:outfile_provider]
+  @keep_outfile = options[:keep_outfile]
 end
 
 def execute_command(obj, options={})
+  timeout = options[:timeout] || 5
+  @busy = false if @busy_start_time && (Time.now > @busy_start_time + timeout)
   if @busy
     do_work
     :backend_busy 
@@ -29,9 +36,10 @@ def execute_command(obj, options={})
     @socket.send(serialize_message(obj), 0)
     result = nil
     @busy = true
+    @busy_start_time = Time.now
     if options[:response_callback]
       @invocations[@invocation_id] = lambda do |r|
-        if r["type"] == "response" || r["type"] =~ /.*error$/
+        if r["type"] == "response" || r["type"] =~ /error$/
           @busy = false
         end
         options[:response_callback].call(r)
@@ -41,15 +49,20 @@ def execute_command(obj, options={})
       :request_pending 
     else
       @invocations[@invocation_id] = lambda do |r|
-        if r["type"] == "response" || r["type"] == /.*error$/ 
+        if r["type"] == "response" || r["type"] =~ /error$/ 
           result = r
           @busy = false
         end
       end
       @invocation_id += 1
       while !result
-        sleep(0.1)
-        do_work
+        if Time.now > @busy_start_time + timeout
+          result = :timeout
+          @busy = false
+        else
+          sleep(0.1)
+          do_work
+        end
       end
       result
     end
@@ -104,14 +117,19 @@ end
 
 def connect
   @state = :connecting
+  @connect_start_time = Time.now
 
   @logger.info @config.command if @logger
 
-  @out_file = tempfile_name 
+  if @outfile_provider
+    @out_file = @outfile_provider.call
+  else
+    @out_file = tempfile_name 
+  end
   File.unlink(@out_file) if File.exist?(@out_file)
 
   Dir.chdir(File.dirname(@config.file)) do
-    @process_id = spawn(@config.command.strip + " > #{@out_file}")
+    @process_id = spawn(@config.command.strip + " > #{@out_file} 2>&1")
   end
   @work_state = :wait_for_file
 end
@@ -122,6 +140,12 @@ def do_work
     if File.exist?(@out_file)
       @work_state = :wait_for_port
     end
+    if Time.now > @connect_start_time + ConnectTimeout
+      cleanup
+      @connection_listener.call(:timeout) if @connection_listener
+      @work_state = :done
+      @state = :off
+    end
     true
   when :wait_for_port
     output = File.read(@out_file)
@@ -131,7 +155,13 @@ def do_work
       @socket = TCPSocket.new("127.0.0.1", port)
       @state = :connected
       @work_state = :read_from_socket
-      @connection_listener.call if @connection_listener
+      @connection_listener.call(:connected) if @connection_listener
+    end
+    if Time.now > @connect_start_time + ConnectTimeout
+      cleanup
+      @connection_listener.call(:timeout) if @connection_listener
+      @work_state = :done
+      @state = :off
     end
     true
   when :read_from_socket
@@ -144,7 +174,7 @@ def do_work
       begin
         data = @socket.read_nonblock(100000)
       rescue Errno::EWOULDBLOCK
-      rescue EOFError
+      rescue IOError, EOFError, Errno::ECONNRESET
         socket_closed = true
         @logger.error "server socket closed (end of file)" if @logger
       end
@@ -162,14 +192,18 @@ def do_work
         end
         true
       elsif !backend_running? || socket_closed
-        @socket.close
-        File.unlink(@out_file)
+        cleanup
         @work_state = :done
         false
       end
     end
   end
 
+end
+
+def cleanup
+  @socket.close if @socket
+  File.unlink(@out_file) unless @keep_outfile
 end
 
 end
